Add Bayesian Layers README.
Currently, the documentation is colocated with the source code. Later, we can think about how to make the documentation easier to find. For example, when we update edwardlib.org to Edward2, it can convert all these README.mds from the source code to display on the website similar to http://edwardlib.org/api/model.

PiperOrigin-RevId: 277088475