Add support for trainable regularizers.
This comes up in variational BNNs when using priors with free parameters. Note this is already possible with Bayesian Layers if you define a free parameter and pass it to the regularizer's constructor arguments:

```python
mean = tf.Variable(tf.random.normal(0.))
regularizer = ed.regularizers.NormalKLDivergence(mean=mean)
```

The advantage of having trainable regularizers is more for self-containment. That is, all trainable parameters are baked into the Keras model and is necessary in practice if you use model.fit instead of a custom training loop.

I added an example regularizer. (The name is a mouthful..)

PiperOrigin-RevId: 280863871