Tune dropout for WRN 28-10 on CIFAR-10 and CIFAR-100.
Results are roughly the same as deterministic on both datasets. For CIFAR-10, other implementations indicate typically a 0.1% accuracy improvement and a slight increase in NLL. It seems like dropout actually worsens performance, with dropout_rate=0.0 doing best; I kept it at 0.1 as we may likely see bigger changes in the OOD evaluations.

PiperOrigin-RevId: 291644260