Fix a bug in the LSTMCellFlipout layer involving incorrect shapes.
In the flipout version of the stochastic LSTM cell, the input sign
vector (as part of the rank-1 reparameterization) had an incorrect
shape of (batch_size, 4 * input_dim). Since the kernel is of shape
(input_dim, 4 * units), the correct shape is (batch_size, input_dim).
Similarly, the recurrent input sign vector had a shape of
(batch_size, 4 * units) when it should be (batch_size, units).

When implementation=2, this bug was readily apparent.

PiperOrigin-RevId: 298650518