Add input_shape type checking for TF1 compatibility.
In TF1, the `input_shape` for the build() method can be tf.Dimension, which is not compatible with the `Identity` initializer (which requires the input dimension to be int rather than tf.Dimension).

This commit fix this by adding an additional type check for input_shape, and convert it to int when appropriate.

PiperOrigin-RevId: 332860116