Updates default scaling scheme for random feature distribution.
When dealing with high-dimensional input from neural networks, it is empirically more effective to use `kernel_init=normal(stddev=sqrt(2/num_input))` and `feature_scale=1.`, instead of `normal(stddev=1.)` and `feature_scale=sqrt(2/num_input)`. Effectively, this means we are applying the sqrt(2/num_input) scaling pre-activation rather than post-activation.

This pull request sets the default parameters to use the post-scaling scheme for random feature layer, so that it works well with neural net model out-of-the-box.

PiperOrigin-RevId: 389699939