Add BatchEnsemble attention and normalization layers.
This is little extra compute for our pretrained models so we should do it for the ensemble capacity+diversity.

This PR only implements the layers. For now, I'm postponing a PR that adds them to vit_batchensemble.py. This will make previous checkpoints that don't do BE on MHA/LN incompatible.

PiperOrigin-RevId: 432125420